backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.11.13
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
blinker==1.9.0
boto3==1.36.8
botocore==1.36.8
certifi==2024.8.30
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crewai==0.102.0
crewai-tools==0.33.0
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.2
fal_client==0.5.9
Flask==3.1.0
Flask-Cors==5.0.1
Flask-Pydantic==0.12.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gotrue==2.11.4
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.3.14
langchain-core==0.3.32
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langgraph==0.3.1
langgraph-checkpoint==2.0.16
langgraph-sdk==0.1.53
langsmith==0.1.147
MarkupSafe==2.1.5
msgpack==1.1.0
multidict==6.1.0
numpy==1.26.3
openai==1.65.3
orjson==3.10.7
packaging==24.1
postgrest==0.19.3
prompt_toolkit==3.0.50
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
realtime==2.4.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
s3transfer==0.11.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.38
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tenacity==9.0.0
tiktoken==0.7.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websockets==14.2
Werkzeug==3.1.3
yarl==1.18.3
zstandard==0.23.0